move shouldTimeout helper, add tests
